cleaned up organization and deleted old documentation.
